CEDS-1056 Fixes for eori and conditional journey
This commit cover changes related with making eori field optional and conditional journey based on the choice.

When user during Departure choose Back To The UK he will be redirected directly to the summary page.
In other case he need to fill two additional screens.

Moreover with this PR I removed old and deprecated CustomsDeclareBaseSpec. Actual MovementBaseSpec contains only the most important stuff.

Tests for MovementSummaryPage was compressed because in the nearest future there will be new implementation of this.
It's easier to fix and maintain less tests for the same page.