Change logging level for unknown notification error codes
There is no such need to have this logging at warn level. It is not that
severe for our service functioning.
Currently it more of an information that we need to investigate.
The other thing is, that errors often comes in different format than it
is written in documentation, so until it is sorted it is better not to
trigger alerts of any kind.