CEDS-1860 Improve algorithm in ViewSubmissionsController
The algorithm is for matching Submissions with Notifications by their
conversation ID.

The time complexity used to be s * n, where:
 - s - number of Submissions
 - n - number of Notifications
For large amounts of those it would get slower quite quickly.
Additionally it was sorting Notifications before the process which was
unnecessary as they were being filtered later.

The new solution makes use of groupBy method and therefore comes down
to logarithmic complexity.