CEDS-2024 Add Label to all input text fields
We have realised that some conditional reveal text input fields do not
have labels. This was also true in the prototype.
Every input text field requires a label.

I have spoken with the product team and they have provided the content
for those labels on version29 of the prototype.

This commit addresses this concern.