CEDS-2028 Migrate Start page to GDS Design system
As part of this work, we kept looking at the prototype. Unfortunately
how the Start page was designed in the prototype it composed of several
`govuk-grid-column-two-thirds` plus some `govuk-grid-column-one-third`.

Since the start page is not a critical page to the service, we have
brought it to the GDS design system, but kept the same layout as before.

Also the Dashed list does not exist as a component in the GDS design
system, so again, we have kept the same approach as before, by simply
using the a normal dash.