CEDS-2028 Add govukFlexibleLayout
The default govukLayout provided by play-frontend-govuk does not allow
for customizable page content width. Everything is encapsulated in
govuk-grid-column-two-thirds element.

The new component is a copy of the library's govukLayout but it is
passing the content as is, without adding wrapping elements.
The responsibility for this is moved to the client.

To choose new layout component you have to set useDefaultContentWidth to
false in gds_main_template.