CEDS-4813 Secure Movements FE login on TDR with additional secret value.
Adding new 'secret.tdrHashSalt' config value that will only be set in the TDR environment.

Adding additional authentication check that if the 'tdrHashSalt' value is present in config, then the user must specify an additional HMRC-CUS-ORG enrolment value of 'TDRSecret' who's value must be a valid hash of the supplied 'EORINumber' value.